+++ pytsite/maintenance/__init__.py
@@ -1,11 +1,11 @@
 """PytSite Maintenance Package.
 """
-# Public API
-from ._api import enable, disable, is_enabled
-
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
+
+# Public API
+from ._api import enable, disable, is_enabled
 
 
 def _init():