+++ pytsite/semver/_api.py
@@ -1,20 +1,20 @@
 """PytSite Semantic Versioning Tools Functions
 """
-import re as _re
-from typing import Optional as _Optional, Union as _Union, Iterable as _Iterable
-from . import _error
-from ._version import Version as _Version
-
-_REQUIREMENT_RE = _re.compile('([a-zA-Z0-9_.\-]+)\s*(==|!=|<=|>=|>|<)?\s*(\d+\.\d+\.\d+|\d+\.\d+|\d+)?')
-_CONDITION_RE = _re.compile('^([<>=!]{1,2})?([0-9\.]+)$')
-_ALLOWED_OPERATORS = ('==', '!=', '>', '<', '>=', '<=')
-
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
+import re as _re
+from typing import Optional as _Optional, Union as _Union, Iterable as _Iterable, Tuple as _Tuple
+from . import _error
+from ._version import Version as _Version
 
-def parse_condition_str(condition: str) -> tuple:
+_REQUIREMENT_RE = _re.compile('([a-zA-Z0-9_.\-]+)\s*(==|!=|<=|>=|>|<)?\s*(\d+\.\d+\.\d+|\d+\.\d+|\d+)?')
+_CONDITION_RE = _re.compile('^([<>=!]{1,2})?([0-9.]+)$')
+_ALLOWED_OPERATORS = ('==', '!=', '>', '<', '>=', '<=')
+
+
+def parse_condition_str(condition: str) -> _Tuple[str, _Version]:
     """Parse a condition string
 
     Condition string must consist of two parts: a comparison operator and a version number,
@@ -32,10 +32,10 @@ def parse_condition_str(condition: str) -> tuple:
         raise _error.InvalidComparisonOperator("Unknown comparison operator '{}' used in condition '{}'".
                                                format(operator, version))
 
-    return operator, version
+    return operator, _Version(version)
 
 
-def parse_requirement_str(requirement: str) -> tuple:
+def parse_requirement_str(requirement: str) -> _Tuple[str, str]:
     """Parse a requirement string
 
     Requirement string is a string contains three parts: name, condition and version, i. e.: 'pytsite > 1.1' or
@@ -51,7 +51,7 @@ def parse_requirement_str(requirement: str) -> tuple:
 
     name, condition, version = match.group(1), match.group(2), match.group(3)
 
-    return name, '{}{}'.format(condition or '>=', version or '0.0.1')
+    return name, '{}{}'.format(condition or '>=', str(_Version(version)) if version else '0.0.1')
 
 
 def parse_version_str(version: str) -> _Version:
@@ -60,14 +60,17 @@ def parse_version_str(version: str) -> _Version:
     return _Version(version)
 
 
-def compare(a: str, b: str) -> int:
+def compare(a: _Union[str, _Version], b: _Union[str, _Version]) -> int:
     """Compare two versions
 
     Returns -1 if a < b, 0 if a == b, 1 if a > b
     """
 
-    a = _Version(a)
-    b = _Version(b)
+    if not isinstance(a, _Version):
+        a = _Version(a)
+
+    if not isinstance(b, _Version):
+        b = _Version(b)
 
     if a < b:
         return -1
@@ -99,7 +102,6 @@ def minimum(condition: str) -> str:
     """Get minimum possible version number for provided condition
     """
     op, ver = parse_condition_str(condition)
-    ver = parse_version_str(ver)
 
     if op == '>':
         return str(ver + 1)
@@ -113,7 +115,6 @@ def maximum(condition: str) -> str:
     """Get maximum possible version number for provided condition
     """
     op, ver = parse_condition_str(condition)
-    ver = parse_version_str(ver)
 
     if op == '<':
         return str(ver - 1)
@@ -144,13 +145,12 @@ def last(versions: _Iterable[str], conditions: _Union[str, _Iterable[str]] = Non
     return filtered[-1] if filtered else None
 
 
-def check_conditions(version_to_check: str, conditions: _Union[str, _Iterable[str]]) -> bool:
+def check_conditions(version_to_check: _Union[str, _Version], conditions: _Union[str, _Iterable[str]]) -> bool:
     """Check if version_to_check satisfies condition
 
     version_to_check should be like '1', '1.0' or '1.0.0'.
     condition should be like '==1', '>=1.0', '!=1.0.0', etc.
     """
-
     if isinstance(conditions, (list, tuple)):
         r = True
         for c in conditions:
@@ -159,25 +159,25 @@ def check_conditions(version_to_check: str, conditions: _Union[str, _Iterable[st
                 return r
         return r
 
-    operator, version_to_comapre = parse_condition_str(conditions)
+    operator, version_to_compare = parse_condition_str(conditions)
 
     if operator == '==':
-        if compare(version_to_check, version_to_comapre) != 0:
+        if compare(version_to_check, version_to_compare) != 0:
             return False
     elif operator == '!=':
-        if compare(version_to_check, version_to_comapre) == 0:
+        if compare(version_to_check, version_to_compare) == 0:
             return False
     elif operator == '<':
-        if compare(version_to_check, version_to_comapre) >= 0:
+        if compare(version_to_check, version_to_compare) >= 0:
             return False
     elif operator == '>':
-        if compare(version_to_check, version_to_comapre) <= 0:
+        if compare(version_to_check, version_to_compare) <= 0:
             return False
     elif operator == '<=':
-        if compare(version_to_check, version_to_comapre) > 0:
+        if compare(version_to_check, version_to_compare) > 0:
             return False
     elif operator == '>=':
-        if compare(version_to_check, version_to_comapre) < 0:
+        if compare(version_to_check, version_to_compare) < 0:
             return False
 
     return True